refactor: extract formatDate, formatTime, displayName to lib/format.ts
